+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # v2.8.7
+
+* added setdefault to `DictObject`
+